+++ src/test/java/group/gnometrading/backtest/exchange/MBPSubmitTest.java
@@ -26,7 +26,7 @@ class MBPSubmitTest {
 
     static class DummyFeeModel implements FeeModel {
         @Override
-        public double calculateFee(long totalPrice, boolean isMaker) {
+        public double calculateFee(double totalPrice, boolean isMaker) {
             return totalPrice * (isMaker ? 0.03 : 0.05);
         }
     }
